Move evaluation logic into featureflag package
